dashboard/requirements.txt
pylint==3.3.8
pytest==8.4.2
pytest-cov==7.0.0
streamlit==1.50.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
requests==2.32.5
bs4==0.0.2